ok bass error fixed (hopefully?)
OK so a guy told me that win7 introduced unified sound cards (which means they use the same buffer) which means that the bass offset post winvista is the SAME on windows. which is like 0.033 (I tested it)

looks like it works, so that's based as hell. WILL BREAK ON LINUX OBV, we can't really do anything there so uh.