weird lua bug
for some reason when using a ptr/unique ptr with sol lua, and you reset it/free it and then make a new one on the same object. It basically fucks itself half of the time on a race condition somewhere, and specifically on usertypes. It will assign the wrong properties.